refactor(watchlist): consolidate mutation hooks into one file to fit PR max-lines limit
Inlines the three per-hook files (add/remove/update) into a single
useTokenWatchlistMutations module sharing a private factory, and merges
their test files into a single parameterized suite, to stay under the
1000-line CI cap. All 45 tests still pass.
