+++ app/components/UI/assets/watchlist/hooks/index.ts
@@ -1,16 +1,12 @@
 export { tokenWatchlistQueryKeys } from './useTokenWatchlist.keys';
 export {
   tokenWatchlistAddBatcher,
-  useTokenWatchlistAddItemMutation,
-  type WatchlistAddInput,
-} from './useTokenWatchlistAddItemMutation';
-export {
   tokenWatchlistRemoveBatcher,
-  useTokenWatchlistRemoveItemMutation,
-  type WatchlistRemoveInput,
-} from './useTokenWatchlistRemoveItemMutation';
-export {
   tokenWatchlistUpdateBatcher,
+  useTokenWatchlistAddItemMutation,
+  useTokenWatchlistRemoveItemMutation,
   useTokenWatchlistUpdateListMutation,
+  type WatchlistAddInput,
+  type WatchlistRemoveInput,
   type WatchlistUpdateListInput,
-} from './useTokenWatchlistUpdateListMutation';
+} from './useTokenWatchlistMutations';
